refactor(JSService): rename service-subject to typed-dependencies
This makes it more clear what the file does.
