+++ src/types/typed-dependencies.type.ts
